requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
numpy==1.22.0
pandas==1.3.5
psycopg==3.0.8
psycopg-binary==3.0.8
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
